+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - Add FreeU for *backend:diffusers*  
   for *backend:original* use extension: <https://github.com/ljleb/sd-webui-freeu>  
 - Add HyperTile: <https://github.com/tfernd/HyperTile>
-- Fix override settings
+- Implement Styles extra field
 
 This is a big one, with some major changes and new functionality...  
 And probably the biggest release since introduction of **Diffusers**